+++ compute/machine_images.go
@@ -7,17 +7,17 @@ type MachineImagesClient struct {
 
 // MachineImages obtains an MachineImagesClient which can be used to access to the
 // MachineImage functions of the Compute API
-func (c *ComputeClient) MachineImages() *MachineImagesClient {
+func (c *Client) MachineImages() *MachineImagesClient {
 	return &MachineImagesClient{
 		ResourceClient: ResourceClient{
-			ComputeClient:       c,
+			Client:              c,
 			ResourceDescription: "MachineImage",
 			ContainerPath:       "/machineimage/",
 			ResourceRootPath:    "/machineimage",
 		}}
 }
 
-// MahineImage describes an existing Machine Image.
+// MachineImage describes an existing Machine Image.
 type MachineImage struct {
 	// account of the associated Object Storage Classic instance
 	Account string `json:"account"`
@@ -117,7 +117,7 @@ func (c *MachineImagesClient) DeleteMachineImage(deleteInput *DeleteMachineImage
 	return c.deleteResource(deleteInput.Name)
 }
 
-// GetMachineList describes the MachineImage to get
+// GetMachineImageInput describes the MachineImage to get
 type GetMachineImageInput struct {
 	// account of the associated Object Storage Classic instance
 	Account string `json:"account"`