+++ compute/orchestration.go
@@ -7,10 +7,10 @@ import (
 	"github.com/hashicorp/go-oracle-terraform/client"
 )
 
-const WaitForOrchestrationActivePollInterval = time.Duration(10 * time.Second)
-const WaitForOrchestrationActiveTimeout = time.Duration(3600 * time.Second)
-const WaitForOrchestrationDeletePollInterval = time.Duration(10 * time.Second)
-const WaitForOrchestrationDeleteTimeout = time.Duration(3600 * time.Second)
+const waitForOrchestrationActivePollInterval = 10 * time.Second
+const waitForOrchestrationActiveTimeout = 3600 * time.Second
+const waitForOrchestrationDeletePollInterval = 10 * time.Second
+const waitForOrchestrationDeleteTimeout = 3600 * time.Second
 
 // OrchestrationsClient is a client for the Orchestration functions of the Compute API.
 type OrchestrationsClient struct {
@@ -19,56 +19,73 @@ type OrchestrationsClient struct {
 
 // Orchestrations obtains an OrchestrationsClient which can be used to access to the
 // Orchestration functions of the Compute API
-func (c *ComputeClient) Orchestrations() *OrchestrationsClient {
+func (c *Client) Orchestrations() *OrchestrationsClient {
 	return &OrchestrationsClient{
 		ResourceClient: ResourceClient{
-			ComputeClient:       c,
+			Client:              c,
 			ResourceDescription: "Orchestration",
 			ContainerPath:       "/platform/v1/orchestration/",
 			ResourceRootPath:    "/platform/v1/orchestration",
 		}}
 }
 
+// OrchestrationDesiredState defines the different desired states a orchestration can be in
 type OrchestrationDesiredState string
 
 const (
-	// * active: Creates all the orchestration objects defined in the orchestration.
+	// OrchestrationDesiredStateActive - Creates all the orchestration objects defined in the orchestration.
 	OrchestrationDesiredStateActive OrchestrationDesiredState = "active"
-	// * inactive: Adds the orchestration to Oracle Compute Cloud Service, but does not create any of the orchestration
+	// OrchestrationDesiredStateInactive - Adds the orchestration to Oracle Compute Cloud Service, but does not create any of the orchestration
 	OrchestrationDesiredStateInactive OrchestrationDesiredState = "inactive"
-	// * suspended: Suspends all orchestration objects defined in the orchestration
+	// OrchestrationDesiredStateSuspend - Suspends all orchestration objects defined in the orchestration
 	OrchestrationDesiredStateSuspend OrchestrationDesiredState = "suspend"
 )
 
+// OrchestrationStatus defines the different status a orchestration can be in
 type OrchestrationStatus string
 
 const (
-	OrchestrationStatusActive       OrchestrationStatus = "active"
-	OrchestrationStatusInactive     OrchestrationStatus = "inactive"
-	OrchestrationStatusSuspend      OrchestrationStatus = "suspend"
-	OrchestrationStatusActivating   OrchestrationStatus = "activating"
-	OrchestrationStatusDeleting     OrchestrationStatus = "deleting"
-	OrchestrationStatusError        OrchestrationStatus = "terminal_error"
-	OrchestrationStatusStopping     OrchestrationStatus = "stopping"
-	OrchestrationStatusSuspending   OrchestrationStatus = "suspending"
-	OrchestrationStatusStarting     OrchestrationStatus = "starting"
+	// OrchestrationStatusActive - active
+	OrchestrationStatusActive OrchestrationStatus = "active"
+	// OrchestrationStatusInactive - inactive
+	OrchestrationStatusInactive OrchestrationStatus = "inactive"
+	// OrchestrationStatusSuspend - suspend
+	OrchestrationStatusSuspend OrchestrationStatus = "suspend"
+	// OrchestrationStatusActivating - activating
+	OrchestrationStatusActivating OrchestrationStatus = "activating"
+	// OrchestrationStatusDeleting - deleting
+	OrchestrationStatusDeleting OrchestrationStatus = "deleting"
+	// OrchestrationStatusError - terminal_error
+	OrchestrationStatusError OrchestrationStatus = "terminal_error"
+	// OrchestrationStatusStopping - stopping
+	OrchestrationStatusStopping OrchestrationStatus = "stopping"
+	// OrchestrationStatusSuspending - suspending
+	OrchestrationStatusSuspending OrchestrationStatus = "suspending"
+	// OrchestrationStatusStarting - starting
+	OrchestrationStatusStarting OrchestrationStatus = "starting"
+	// OrchestrationStatusDeactivating - deactivating
 	OrchestrationStatusDeactivating OrchestrationStatus = "deactivating"
-	OrchestrationStatusSuspended    OrchestrationStatus = "suspended"
+	// OrchestrationStatusSuspended - suspended
+	OrchestrationStatusSuspended OrchestrationStatus = "suspended"
 )
 
+// OrchestrationType defines the type of orchestrations that can be managed
 type OrchestrationType string
 
 const (
+	// OrchestrationTypeInstance - Instance
 	OrchestrationTypeInstance OrchestrationType = "Instance"
 )
 
+// OrchestrationRelationshipType defines the orchestration relationship type for an orchestration
 type OrchestrationRelationshipType string
 
 const (
+	// OrchestrationRelationshipTypeDepends - the orchestration relationship depends on a resource
 	OrchestrationRelationshipTypeDepends OrchestrationRelationshipType = "depends"
 )
 
-// OrchestrationInfo describes an existing Orchestration.
+// Orchestration describes an existing Orchestration.
 type Orchestration struct {
 	// The default Oracle Compute Cloud Service account, such as /Compute-acme/default.
 	Account string `json:"account"`
@@ -133,6 +150,7 @@ type CreateOrchestrationInput struct {
 	Timeout time.Duration `json:"-"`
 }
 
+// Object defines an object inside an orchestration
 type Object struct {
 	// The default Oracle Compute Cloud Service account, such as /Compute-acme/default.
 	// Optional
@@ -192,6 +210,7 @@ type Object struct {
 	Version int `json:"version,omitempty"`
 }
 
+// Health defines the health of an object
 type Health struct {
 	// The status of the object
 	Status OrchestrationStatus `json:"status,omitempty"`
@@ -203,6 +222,7 @@ type Health struct {
 	Error string `json:"error,omitempty"`
 }
 
+// Relationship defines the relationship between objects
 type Relationship struct {
 	// The type of Relationship
 	// The only type is depends
@@ -221,7 +241,7 @@ func (c *OrchestrationsClient) CreateOrchestration(input *CreateOrchestrationInp
 	for _, i := range input.Objects {
 		i.Orchestration = c.getQualifiedName(i.Orchestration)
 		if i.Type == OrchestrationTypeInstance {
-			instanceClient := c.ComputeClient.Instances()
+			instanceClient := c.Client.Instances()
 			instanceInput := i.Template.(*CreateInstanceInput)
 			instanceInput.Name = c.getQualifiedName(instanceInput.Name)
 
@@ -256,10 +276,10 @@ func (c *OrchestrationsClient) CreateOrchestration(input *CreateOrchestrationInp
 	}
 
 	if input.PollInterval == 0 {
-		input.PollInterval = WaitForOrchestrationActivePollInterval
+		input.PollInterval = waitForOrchestrationActivePollInterval
 	}
 	if input.Timeout == 0 {
-		input.Timeout = WaitForOrchestrationActiveTimeout
+		input.Timeout = waitForOrchestrationActiveTimeout
 	}
 
 	// Wait for orchestration to be ready and return the result
@@ -350,10 +370,10 @@ func (c *OrchestrationsClient) UpdateOrchestration(input *UpdateOrchestrationInp
 	}
 
 	if input.PollInterval == 0 {
-		input.PollInterval = WaitForOrchestrationActivePollInterval
+		input.PollInterval = waitForOrchestrationActivePollInterval
 	}
 	if input.Timeout == 0 {
-		input.Timeout = WaitForOrchestrationActiveTimeout
+		input.Timeout = waitForOrchestrationActiveTimeout
 	}
 
 	// Wait for orchestration to be ready and return the result
@@ -384,10 +404,10 @@ func (c *OrchestrationsClient) DeleteOrchestration(input *DeleteOrchestrationInp
 	}
 
 	if input.PollInterval == 0 {
-		input.PollInterval = WaitForOrchestrationDeletePollInterval
+		input.PollInterval = waitForOrchestrationDeletePollInterval
 	}
 	if input.Timeout == 0 {
-		input.Timeout = WaitForOrchestrationDeleteTimeout
+		input.Timeout = waitForOrchestrationDeleteTimeout
 	}
 
 	return c.WaitForOrchestrationDeleted(input, input.PollInterval, input.Timeout)
@@ -397,7 +417,7 @@ func (c *OrchestrationsClient) success(info *Orchestration) (*Orchestration, err
 	c.unqualify(&info.Name)
 	for _, i := range info.Objects {
 		c.unqualify(&i.Orchestration)
-		if OrchestrationType(i.Type) == OrchestrationTypeInstance {
+		if i.Type == OrchestrationTypeInstance {
 			instanceInput := i.Template.(map[string]interface{})
 			instanceInput["name"] = c.getUnqualifiedName(instanceInput["name"].(string))
 		}
@@ -406,7 +426,7 @@ func (c *OrchestrationsClient) success(info *Orchestration) (*Orchestration, err
 	return info, nil
 }
 
-// WaitForOrchestrationActive waits for an orchestration to be completely initialized and available.
+// WaitForOrchestrationState waits for an orchestration to be in the specified state
 func (c *OrchestrationsClient) WaitForOrchestrationState(input *GetOrchestrationInput, pollInterval, timeout time.Duration) (Orchestration, error) {
 	var info *Orchestration
 	var getErr error
@@ -444,9 +464,8 @@ func (c *OrchestrationsClient) WaitForOrchestrationState(input *GetOrchestration
 			c.client.DebugLogString("Orchestration suspended")
 			if info.DesiredState == OrchestrationDesiredStateSuspend {
 				return true, nil
-			} else {
-				return false, nil
 			}
+			return false, nil
 		default:
 			return false, fmt.Errorf("Unknown orchestration state: %s, erroring", s)
 		}