+++ compute/security_lists.go
@@ -7,21 +7,25 @@ type SecurityListsClient struct {
 
 // SecurityLists obtains a SecurityListsClient which can be used to access to the
 // Security List functions of the Compute API
-func (c *ComputeClient) SecurityLists() *SecurityListsClient {
+func (c *Client) SecurityLists() *SecurityListsClient {
 	return &SecurityListsClient{
 		ResourceClient: ResourceClient{
-			ComputeClient:       c,
+			Client:              c,
 			ResourceDescription: "security list",
 			ContainerPath:       "/seclist/",
 			ResourceRootPath:    "/seclist",
 		}}
 }
 
+// SecurityListPolicy defines the constants a security list policy can be
 type SecurityListPolicy string
 
 const (
-	SecurityListPolicyDeny   SecurityListPolicy = "deny"
+	// SecurityListPolicyDeny - deny
+	SecurityListPolicyDeny SecurityListPolicy = "deny"
+	// SecurityListPolicyReject - reject
 	SecurityListPolicyReject SecurityListPolicy = "reject"
+	// SecurityListPolicyPermit - permit
 	SecurityListPolicyPermit SecurityListPolicy = "permit"
 )