+++ compute/storage_volumes.go
@@ -9,10 +9,10 @@ import (
 	"github.com/hashicorp/go-oracle-terraform/client"
 )
 
-const WaitForVolumeReadyPollInterval = time.Duration(10 * time.Second)
-const WaitForVolumeReadyTimeout = time.Duration(600 * time.Second)
-const WaitForVolumeDeletePollInterval = time.Duration(10 * time.Second)
-const WaitForVolumeDeleteTimeout = time.Duration(600 * time.Second)
+const waitForVolumeReadyPollInterval = 10 * time.Second
+const waitForVolumeReadyTimeout = 600 * time.Second
+const waitForVolumeDeletePollInterval = 10 * time.Second
+const waitForVolumeDeleteTimeout = 600 * time.Second
 
 // StorageVolumeClient is a client for the Storage Volume functions of the Compute API.
 type StorageVolumeClient struct {
@@ -21,10 +21,10 @@ type StorageVolumeClient struct {
 
 // StorageVolumes obtains a StorageVolumeClient which can be used to access to the
 // Storage Volume functions of the Compute API
-func (c *ComputeClient) StorageVolumes() *StorageVolumeClient {
+func (c *Client) StorageVolumes() *StorageVolumeClient {
 	return &StorageVolumeClient{
 		ResourceClient: ResourceClient{
-			ComputeClient:       c,
+			Client:              c,
 			ResourceDescription: "storage volume",
 			ContainerPath:       "/storage/volume/",
 			ResourceRootPath:    "/storage/volume",
@@ -32,22 +32,22 @@ func (c *ComputeClient) StorageVolumes() *StorageVolumeClient {
 
 }
 
+// StorageVolumeKind defines the kinds of storage volumes that can be managed
 type StorageVolumeKind string
 
 const (
+	// StorageVolumeKindDefault - "/oracle/public/storage/default"
 	StorageVolumeKindDefault StorageVolumeKind = "/oracle/public/storage/default"
+	// StorageVolumeKindLatency - "/oracle/public/storage/latency"
 	StorageVolumeKindLatency StorageVolumeKind = "/oracle/public/storage/latency"
-	StorageVolumeKindSSD     StorageVolumeKind = "/oracle/public/storage/ssd/gpl"
+	// StorageVolumeKindSSD - "/oracle/public/storage/ssd/gpl"
+	StorageVolumeKindSSD StorageVolumeKind = "/oracle/public/storage/ssd/gpl"
 )
 
 // StorageVolumeInfo represents information retrieved from the service about a Storage Volume.
 type StorageVolumeInfo struct {
 	// Shows the default account for your identity domain.
 	Account string `json:"account,omitempty"`
-
-	// true indicates that the storage volume can also be used as a boot disk for an instance.
-	// If you set the value to true, then you must specify values for the `ImageList` and `ImageListEntry` fields.
-	Bootable bool `json:"bootable,omitempty"`
 
 	// The description of the storage volume.
 	Description string `json:"description,omitempty"`
@@ -58,14 +58,8 @@ type StorageVolumeInfo struct {
 	// Name of machine image to extract onto this volume when created. This information is provided only for bootable storage volumes.
 	ImageList string `json:"imagelist,omitempty"`
 
-	// Specific imagelist entry version to extract.
-	ImageListEntry int `json:"imagelist_entry,omitempty"`
-
 	// Three-part name of the machine image. This information is available if the volume is a bootable storage volume.
 	MachineImage string `json:"machineimage_name,omitempty"`
-
-	// All volumes are managed volumes. Default value is true.
-	Managed bool `json:"managed,omitempty"`
 
 	// The three-part name of the object (/Compute-identity_domain/user/object).
 	Name string `json:"name"`
@@ -75,9 +69,6 @@ type StorageVolumeInfo struct {
 
 	// The storage-pool property: /oracle/public/storage/latency or /oracle/public/storage/default.
 	Properties []string `json:"properties,omitempty"`
-
-	// Boolean field indicating whether this volume can be attached as readonly. If set to False the volume will be attached as read-write.
-	ReadOnly bool `json:"readonly,omitempty"`
 
 	// The size of this storage volume in GB.
 	Size string `json:"size"`
@@ -109,6 +100,19 @@ type StorageVolumeInfo struct {
 
 	// Uniform Resource Identifier
 	URI string `json:"uri,omitempty"`
+
+	// true indicates that the storage volume can also be used as a boot disk for an instance.
+	// If you set the value to true, then you must specify values for the `ImageList` and `ImageListEntry` fields.
+	Bootable bool `json:"bootable,omitempty"`
+
+	// All volumes are managed volumes. Default value is true.
+	Managed bool `json:"managed,omitempty"`
+
+	// Boolean field indicating whether this volume can be attached as readonly. If set to False the volume will be attached as read-write.
+	ReadOnly bool `json:"readonly,omitempty"`
+
+	// Specific imagelist entry version to extract.
+	ImageListEntry int `json:"imagelist_entry,omitempty"`
 }
 
 func (c *StorageVolumeClient) getStorageVolumePath(name string) string {
@@ -170,16 +174,16 @@ func (c *StorageVolumeClient) CreateStorageVolume(input *CreateStorageVolumeInpu
 	input.Size = sizeInBytes
 
 	var storageInfo StorageVolumeInfo
-	if err := c.createResource(&input, &storageInfo); err != nil {
+	if err = c.createResource(&input, &storageInfo); err != nil {
 		return nil, err
 	}
 
 	// Should never be nil, as we set this in the provider; but protect against it anyways.
 	if input.PollInterval == 0 {
-		input.PollInterval = WaitForVolumeReadyPollInterval
+		input.PollInterval = waitForVolumeReadyPollInterval
 	}
 	if input.Timeout == 0 {
-		input.Timeout = WaitForVolumeReadyTimeout
+		input.Timeout = waitForVolumeReadyTimeout
 	}
 
 	volume, err := c.waitForStorageVolumeToBecomeAvailable(input.Name, input.PollInterval, input.Timeout)
@@ -189,7 +193,7 @@ func (c *StorageVolumeClient) CreateStorageVolume(input *CreateStorageVolumeInpu
 				Name: volume.Name,
 			}
 
-			if err := c.DeleteStorageVolume(deleteInput); err != nil {
+			if err = c.DeleteStorageVolume(deleteInput); err != nil {
 				return nil, err
 			}
 		}
@@ -217,10 +221,10 @@ func (c *StorageVolumeClient) DeleteStorageVolume(input *DeleteStorageVolumeInpu
 
 	// Should never be nil, but protect against it anyways
 	if input.PollInterval == 0 {
-		input.PollInterval = WaitForVolumeDeletePollInterval
+		input.PollInterval = waitForVolumeDeletePollInterval
 	}
 	if input.Timeout == 0 {
-		input.Timeout = WaitForVolumeDeleteTimeout
+		input.Timeout = waitForVolumeDeleteTimeout
 	}
 
 	return c.waitForStorageVolumeToBeDeleted(input.Name, input.PollInterval, input.Timeout)
@@ -317,10 +321,10 @@ func (c *StorageVolumeClient) UpdateStorageVolume(input *UpdateStorageVolumeInpu
 	}
 
 	if input.PollInterval == 0 {
-		input.PollInterval = WaitForVolumeReadyPollInterval
+		input.PollInterval = waitForVolumeReadyPollInterval
 	}
 	if input.Timeout == 0 {
-		input.Timeout = WaitForVolumeReadyTimeout
+		input.Timeout = waitForVolumeReadyTimeout
 	}
 
 	volumeInfo, err := c.waitForStorageVolumeToBecomeAvailable(input.Name, input.PollInterval, input.Timeout)