+++ database/access_rules_test.go
@@ -2,6 +2,7 @@ package database
 
 import (
 	"fmt"
+	"math/rand"
 	"testing"
 
 	"os"
@@ -20,8 +21,6 @@ const (
 	_TestAccessRuleSource      = "PUBLIC-INTERNET"
 )
 
-var _TestAccessRuleName = fmt.Sprintf("test-acc-rule-%d", helper.RInt())
-
 func TestAccAccessRulesLifeCycle(t *testing.T) {
 	helper.Test(t, helper.TestCase{})
 
@@ -30,10 +29,13 @@ func TestAccAccessRulesLifeCycle(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	var instanceName string
+	var (
+		instanceName string
+		sInstance    *ServiceInstance
+	)
 	if v := os.Getenv("OPC_TEST_DB_INSTANCE"); v == "" {
 		// First Create a Service Instance
-		sInstance, err := sClient.createTestServiceInstance()
+		sInstance, err = sClient.createTestServiceInstance()
 		if err != nil {
 			t.Fatalf("Error creating Service Instance: %s", err)
 		}
@@ -44,13 +46,16 @@ func TestAccAccessRulesLifeCycle(t *testing.T) {
 		instanceName = v
 	}
 
+	rInt := rand.Int()
+	testAccessRuleName := fmt.Sprintf("test-acc-rule-%d", rInt)
+
 	// Create an Access Rule that's disabled
 	input := &CreateAccessRuleInput{
 		ServiceInstanceID: instanceName,
 		Description:       _TestAccessRuleDescription,
 		Destination:       AccessRuleDefaultDestination,
 		Ports:             _TestAccessRulePorts,
-		Name:              _TestAccessRuleName,
+		Name:              testAccessRuleName,
 		Source:            _TestAccessRuleSource,
 		Status:            AccessRuleDisabled,
 	}
@@ -59,26 +64,28 @@ func TestAccAccessRulesLifeCycle(t *testing.T) {
 		Description: _TestAccessRuleDescription,
 		Destination: AccessRuleDefaultDestination,
 		Ports:       _TestAccessRulePorts,
-		Name:        _TestAccessRuleName,
+		Name:        testAccessRuleName,
 		Source:      _TestAccessRuleSource,
 		Status:      AccessRuleDisabled,
 		RuleType:    AccessRuleTypeUser,
 	}
 
 	// Create Access Rule
-	if _, err := aClient.CreateAccessRule(input); err != nil {
+	if _, err = aClient.CreateAccessRule(input); err != nil {
 		t.Fatalf("Error creating AccessRule: %s", err)
 	}
-	defer destroyAccessRule(t, aClient, instanceName, _TestAccessRuleName)
+	defer destroyAccessRule(t, aClient, instanceName, testAccessRuleName)
 
 	// Get Access Rule (Create only returns AccessRule name)
 	getInput := &GetAccessRuleInput{
 		ServiceInstanceID: instanceName,
-		Name:              _TestAccessRuleName,
+		Name:              testAccessRuleName,
 	}
 
-	// Read Result
-	result, err := aClient.GetAccessRule(getInput)
+	var (
+		result *AccessRuleInfo
+	)
+	result, err = aClient.GetAccessRule(getInput)
 	if err != nil {
 		t.Fatalf("Error reading AccessRule: %s", err)
 	}
@@ -91,11 +98,11 @@ func TestAccAccessRulesLifeCycle(t *testing.T) {
 	// Update Access Rule
 	updateInput := &UpdateAccessRuleInput{
 		ServiceInstanceID: instanceName,
-		Name:              _TestAccessRuleName,
+		Name:              testAccessRuleName,
 		Status:            AccessRuleEnabled,
 	}
 
-	if _, err := aClient.UpdateAccessRule(updateInput); err != nil {
+	if _, err = aClient.UpdateAccessRule(updateInput); err != nil {
 		t.Fatalf("Error updating AccessRule: %s", err)
 	}
 
@@ -120,10 +127,13 @@ func TestGetDefaultAccessRule_Basic(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	var instanceName string
+	var (
+		instanceName string
+		sInstance    *ServiceInstance
+	)
 	if v := os.Getenv("OPC_TEST_DB_INSTANCE"); v == "" {
 		// First Create a Service Instance
-		sInstance, err := sClient.createTestServiceInstance()
+		sInstance, err = sClient.createTestServiceInstance()
 		if err != nil {
 			t.Fatalf("Error creating Service Instance: %s", err)
 		}
@@ -163,10 +173,13 @@ func TestUpdateDefaultAccessRule_Basic(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	var instanceName string
+	var (
+		instanceName string
+		sInstance    *ServiceInstance
+	)
 	if v := os.Getenv("OPC_TEST_DB_INSTANCE"); v == "" {
 		// First Create a Service Instance
-		sInstance, err := sClient.createTestServiceInstance()
+		sInstance, err = sClient.createTestServiceInstance()
 		if err != nil {
 			t.Fatalf("Error creating Service Instance: %s", err)
 		}
@@ -224,8 +237,9 @@ func (c *ServiceInstanceClient) createTestServiceInstance() (*ServiceInstance, e
 		UsableStorage:     _ServiceInstanceUsableStorage,
 	}
 
+	rInt := rand.Int()
 	createServiceInstance := &CreateServiceInstanceInput{
-		Name:             fmt.Sprintf("test-acc-instance-%d", helper.RInt()),
+		Name:             fmt.Sprintf("test-acc-instance-%d", rInt),
 		Edition:          _ServiceInstanceEdition,
 		Level:            _ServiceInstanceLevel,
 		Shape:            _ServiceInstanceShape,